requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
requests
docker
python-dotenv
